AutoSpawner # Upgrade and convert to BasicHack
